l3cfg: add support for IPv4 link local addresses (ipv4ll) to NML3Cfg
NML3Cfg already handles IPv4 ACD. IPv4LL is just a small additional
layer on top of that, so it makes sense that it also is handled by
NML3Cfg.

Also, the overall goal is that multiple NMDevice and NMVpnConnection
instances can cooperate independenly. So if one of the "users" enables
IPv4LL on an interface, then the responsible NML3Cfg instance manages
it.

Also, we thus will no longer need to use sd_ipv4ll/n-ipv4ll, because
we implement it ourself.

Also, maybe the code would be "simpler" by not having it part of
NML3Cfg. That would be possible (and still will be in the future),
but since ACD/IPv4LL is basically lot of state, it's not clear
that managing the state becomes simpler by spreading it out. As
it is, it's all side by side.